refactor: templates and test updates
