+++ src/Sinch/SMS/Batches/Send/BatchBase.cs
@@ -105,6 +105,8 @@ namespace Sinch.SMS.Batches.Send
 
         /// <summary>
         ///     Override the default event destination URL for this batch. Must be a valid URL.
+        ///     Learn how to set a default event destination target URL
+        ///     <see href="https://community.sinch.com/t5/SMS/How-do-I-assign-a-callback-URL-to-an-SMS-service-plan/ta-p/8414">here</see>
         /// </summary>
         [JsonPropertyName("callback_url")]
         public Uri? EventDestinationTarget { get; set; }
